// SDKs/twilio.com/twilio_autopilot_v1/1.37.3/typescript/src/sdk/models/shared/autopilotv1assistantfieldtypefieldvalue.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class AutopilotV1AssistantFieldTypeFieldValue extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=assistant_sid" })
  assistantSid?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=field_type_sid" })
  fieldTypeSid?: string;

  @SpeakeasyMetadata({ data: "json, name=language" })
  language?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=synonym_of" })
  synonymOf?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;

  @SpeakeasyMetadata({ data: "json, name=value" })
  value?: string;
}